+++ test/https_test.dart
@@ -9,7 +9,10 @@ void main() {
     test('happy path integration test', () async {
       var host = Platform.environment['FIREBASE_HTTP_BASE_URL'];
       var response = await http.get('$host/helloWorld');
-      expect(response.body, 'Hello from Dart Firebase Functions interop');
+      expect(
+          response.body,
+          'Hello from Dart Firebase Functions interop. '
+          'Here is my secret config: 123456, https://example.com');
     });
   });
 }