+++ tool/deploy.sh
@@ -15,6 +15,8 @@ if [ -z "$TRAVIS" ]; then
     cp build/bin/clear.dart.js index.js
     firebase deploy --only functions
 
+    firebase functions:config:set someservice.key=123456 someservice.url="https://example.com"
+
     # Create new functions
     cp build/bin/index.dart.js index.js
     firebase deploy --only functions
@@ -24,6 +26,8 @@ else
     # The only way I know to delete existing functions is to deploy empty module:
     cp build/bin/clear.dart.js index.js
     firebase deploy --only functions --token "$FIREBASE_TOKEN" --project "$FIREBASE_PROJECT_ID"
+
+    firebase functions:config:set --token="$FIREBASE_TOKEN" someservice.key=123456 someservice.url="https://example.com"
 
     # Create new functions
     cp build/bin/index.dart.js index.js